[discretizers] merge finite volume discretizers
