+++ src/pymor/discretizers/elliptic.py
@@ -146,117 +146,3 @@ def discretize_elliptic_cg(analytical_problem, diameter=None, domain_discretizer
                                               parameter_space=parameter_space, name='{}_CG'.format(p.name))
 
     return discretization, {'grid': grid, 'boundary_info': boundary_info}
-
-
-def discretize_elliptic_fv(analytical_problem, diameter=None, domain_discretizer=None,
-                           grid=None, boundary_info=None):
-    """Discretizes an |EllipticProblem| using the finite volume method.
-
-    Parameters
-    ----------
-    analytical_problem
-        The |EllipticProblem| to discretize.
-    diameter
-        If not `None`, `diameter` is passed as an argument to the `domain_discretizer`.
-    domain_discretizer
-        Discretizer to be used for discretizing the analytical domain. This has
-        to be a function `domain_discretizer(domain_description, diameter, ...)`.
-        If `None`, |discretize_domain_default| is used.
-    grid
-        Instead of using a domain discretizer, the |Grid| can also be passed directly
-        using this parameter.
-    boundary_info
-        A |BoundaryInfo| specifying the boundary types of the grid boundary entities.
-        Must be provided if `grid` is specified.
-
-    Returns
-    -------
-    discretization
-        The |Discretization| that has been generated.
-    data
-        Dictionary with the following entries:
-
-            :grid:           The generated |Grid|.
-            :boundary_info:  The generated |BoundaryInfo|.
-    """
-
-    assert isinstance(analytical_problem, EllipticProblem)
-    assert grid is None or boundary_info is not None
-    assert boundary_info is None or grid is not None
-    assert grid is None or domain_discretizer is None
-
-    p = analytical_problem
-
-    if not (p.nonlinear_advection == p.nonlinear_advection_derivative == p.nonlinear_reaction == None):
-        raise NotImplementedError
-    if analytical_problem.robin_data is not None:
-        raise NotImplementedError
-
-    if grid is None:
-        domain_discretizer = domain_discretizer or discretize_domain_default
-        if diameter is None:
-            grid, boundary_info = domain_discretizer(analytical_problem.domain)
-        else:
-            grid, boundary_info = domain_discretizer(analytical_problem.domain, diameter=diameter)
-
-    Li, coefficients = [], []
-
-    # diffusion part
-    if isinstance(p.diffusion, LincombFunction):
-        Li += [fv.DiffusionOperator(grid, boundary_info, diffusion_function=df, name='diffusion_{}'.format(i))
-               for i, df in enumerate(p.diffusion.functions)]
-        coefficients += p.diffusion.coefficients
-    elif p.diffusion is not None:
-        Li += [fv.DiffusionOperator(grid, boundary_info, diffusion_function=p.diffusion, name='diffusion')]
-        coefficients += [1.]
-
-    # advection part
-    if isinstance(p.advection, LincombFunction):
-        Li += [fv.LinearAdvectionLaxFriedrichs(grid, boundary_info, af, name='advection_{}'.format(i))
-               for i, af in enumerate(p.advection.functions)]
-        coefficients += list(p.advection.coefficients)
-    elif p.advection is not None:
-        Li += [fv.LinearAdvectionLaxFriedrichs(grid, boundary_info, p.advection, name='advection')]
-        coefficients.append(1.)
-
-    # reaction part
-    if isinstance(p.reaction, LincombFunction):
-        raise NotImplementedError
-    elif p.reaction is not None:
-        Li += [fv.ReactionOperator(grid, p.reaction, name='reaction')]
-        coefficients += [1.]
-
-    # system operator
-    if len(coefficients) == 1 and coefficients[0] == 1.:
-        L = Li[0]
-    else:
-        L = LincombOperator(operators=Li, coefficients=coefficients, name='elliptic_operator')
-
-    # rhs
-    if p.dirichlet_data is not None and isinstance(p.diffusion, LincombFunction):
-        F0 = fv.L2ProductFunctional(grid, p.rhs, boundary_info=boundary_info, neumann_data=p.neumann_data)
-        Fi = [fv.L2ProductFunctional(grid, None, boundary_info=boundary_info, dirichlet_data=p.dirichlet_data,
-                                     diffusion_function=df, name='dirichlet_{}'.format(i))
-              for i, df in enumerate(p.diffusion.functions)]
-        F = LincombOperator(operators=[F0] + Fi, coefficients=[1.] + list(p.diffusion.coefficients),
-                            name='rhs')
-    else:
-        F = fv.L2ProductFunctional(grid, p.rhs, boundary_info=boundary_info, dirichlet_data=p.dirichlet_data,
-                                   diffusion_function=p.diffusion, neumann_data=p.neumann_data)
-
-    if grid.reference_element in (triangle, square):
-        visualizer = PatchVisualizer(grid=grid, bounding_box=grid.bounding_box(), codim=0)
-    elif grid.reference_element is line:
-        visualizer = Matplotlib1DVisualizer(grid=grid, codim=0)
-    else:
-        visualizer = None
-
-    l2_product = fv.L2Product(grid, name='l2')
-    products = {'l2': l2_product}
-
-    parameter_space = p.parameter_space if hasattr(p, 'parameter_space') else None
-
-    discretization = StationaryDiscretization(L, F, products=products, visualizer=visualizer,
-                                              parameter_space=parameter_space, name='{}_FV'.format(p.name))
-
-    return discretization, {'grid': grid, 'boundary_info': boundary_info}
