+++ src/pymordemos/elliptic_oned.py
@@ -22,7 +22,8 @@ Options:
 from docopt import docopt
 
 from pymor.analyticalproblems.elliptic import EllipticProblem
-from pymor.discretizers.elliptic import discretize_elliptic_cg, discretize_elliptic_fv
+from pymor.discretizers.elliptic import discretize_elliptic_cg
+from pymor.discretizers.fv import discretize_stationary_fv
 from pymor.domaindescriptions.basic import LineDomain
 from pymor.functions.basic import ExpressionFunction, ConstantFunction, LincombFunction
 from pymor.parameters.functionals import ProjectionParameterFunctional, ExpressionParameterFunctional
@@ -57,7 +58,7 @@ def elliptic_oned_demo(args):
     )
 
     print('Discretize ...')
-    discretizer = discretize_elliptic_fv if args['--fv'] else discretize_elliptic_cg
+    discretizer = discretize_stationary_fv if args['--fv'] else discretize_elliptic_cg
     discretization, _ = discretizer(problem, diameter=1 / args['N'])
 
     print('The parameter type is {}'.format(discretization.parameter_type))
