+++ src/pymortests/fixtures/discretization.py
@@ -7,7 +7,7 @@ from itertools import product
 import pytest
 from pkg_resources import resource_filename
 
-from pymor.discretizers.advection import discretize_nonlinear_instationary_advection_fv
+from pymor.discretizers.fv import discretize_instationary_fv
 from pymor.discretizers.disk import discretize_stationary_from_disk, discretize_instationary_from_disk
 from pymor.discretizers.elliptic import discretize_elliptic_cg
 from pymortests.fixtures.analyticalproblem import (picklable_thermalblock_problems, non_picklable_thermalblock_problems,
@@ -17,7 +17,7 @@ from pymortests.fixtures.analyticalproblem import (picklable_thermalblock_proble
 picklable_discretizaion_generators = \
         [lambda p=p, d=d: discretize_elliptic_cg(p, diameter=d)[0]
          for p, d in product(picklable_thermalblock_problems, [1./50., 1./100.])] + \
-        [lambda p=p, d=d: discretize_nonlinear_instationary_advection_fv(p, diameter=d)[0]
+        [lambda p=p, d=d: discretize_instationary_fv(p, diameter=d)[0]
          for p, d in product(burgers_problems, [1./10., 1./15.])] + \
         [lambda p=p: discretize_stationary_from_disk(parameter_file=p)
          for p in (resource_filename('pymortests', 'testdata/parameter_stationary.ini'),)] + \
